# docs/bulvar-openapi.yaml
openapi: 3.0.3
info:
  title: Bulvar Operational API
  version: 1.0.0
  description: >
    OpenAPI contract for the Bulvar operational surface. The owner chain is
    Poster -> Supabase views/RPC -> ERP UI. The API must not merge Poster
    leftovers into the read model or recalculate min-stock in the child layer.
servers:
  - url: /api
...
paths:
  /bulvar/orders:
    get:
      tags: [Bulvar]
      summary: Load the Bulvar operational matrix rows
      description: Returns the owner-backed read model for the Bulvar dashboard.
      responses:
        '200':
          description: Bulvar order rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BulvarOrderRow'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /bulvar/production-detail:
    get:
      tags: [Bulvar]
      summary: Read Bulvar production detail
      responses:
        '200':
          description: Production rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductionDetailRow'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /bulvar/distribution/run:
    post:
      tags: [Bulvar]
      summary: Rebuild the daily Bulvar distribution result set
      description: >
        Syncs the upstream production snapshot and then runs the Supabase owner
        RPC. The runner writes only to distribution_results and returns the
        batch summary.
      responses:
        '200':
          description: Distribution run result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DistributionRunResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Distribution is already running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
  /bulvar/distribution/results:
    get:
      tags: [Bulvar]
      summary: Read the current Bulvar distribution results
      description: Returns the persisted rows for the current Kyiv business date.
      responses:
        '200':
          description: Distribution result rows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DistributionResultRow'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  responses:
    Unauthorized:
      description: Unauthorized
...
    ServerError:
      description: Server error
  schemas:
    BulvarOrderRow:
      type: object
      additionalProperties: true
      required: [product_id, product_name, spot_name, stock_now, min_stock, avg_sales_day, need_net]
      properties:
        product_id:
          type: integer
        product_name:
          type: string
        spot_name:
          type: string
        store_id:
          type: integer
        stock_now:
          type: number
        min_stock:
          type: number
        avg_sales_day:
          type: number
        need_net:
          type: number
        baked_at_factory:
          type: number
        unit:
          type: string
...
    ProductionDetailRow:
      type: object
      additionalProperties: true
      properties:
        product_id:
          type: integer
        product_name:
          type: string
        baked_at_factory:
          type: number
        total_qty_180d:
          type: number
        prod_days:
          type: number
        avg_qty_per_prod_day:
          type: number
        last_manufacture_at:
          type: string
          nullable: true
        stock_now:
          type: number
        min_stock:
          type: number
        avg_sales_day:
          type: number
        need_net:
          type: number
...
    DistributionRunResponse:
      type: object
      additionalProperties: true
      properties:
        success:
          type: boolean
        batch_id:
          type: string
        products_processed:
          type: integer
        total_kg:
          type: number
        message:
          type: string
    DistributionResultRow:
      type: object
      additionalProperties: true
      properties:
        product_name:
          type: string
        spot_name:
          type: string
        quantity_to_ship:
          type: number
        calculation_batch_id:
          type: string
        business_date:
          type: string
          format: date
        delivery_status:
          type: string
        calc_time:
          type: string
          format: date-time
...